requirements.txt
streamlit==1.38.0
sentence-transformers==2.7.0
chromadb==0.5.5
pypdf==5.0.1
python-docx==1.1.2
openai==1.43.0
tiktoken==0.7.0
